% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key_delete}
\alias{key_delete}
\title{Delete a key from a keyring}
\usage{
key_delete(keyring, key)
}
\arguments{
\item{keyring}{character(1); Name of keyring}

\item{key}{character(1); Name of key}
}
\value{
logical(1); Success of operation
}
\description{
Delete a key from an unlocked keyring.
}
\examples{
\dontrun{
key_delete('mypersonalkeyring', 'key1')
}

}
