% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key_exists}
\alias{key_exists}
\title{Does a given key exist in a keyring}
\usage{
key_exists(keyring, key)
}
\arguments{
\item{keyring}{character(1); Name of keyring}

\item{key}{character(1); Name of key}
}
\value{
logical(1); Existance of key in keyring
}
\description{
In an unlocked keyring return if a key exists.
}
\examples{
\dontrun{
key_exists('mypersonalkeyring', 'key1')
}
}
