% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key_list}
\alias{key_list}
\title{Returns vector of keys in a keyring.}
\usage{
key_list(keyring)
}
\arguments{
\item{keyring}{character(1); Name of keyring}
}
\value{
character; Key names
}
\description{
Return vector key names in a keyring that is unlocked.
}
