% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key_set}
\alias{key_set}
\title{Set a key secret in a keyring}
\usage{
key_set(keyring, key, secret)
}
\arguments{
\item{keyring}{character(1); Name of keyring}

\item{key}{character(1); Name of key to store in keyring}

\item{secret}{character(1); The secret to store in keyring}
}
\value{
logical(1); Status of operation
}
\description{
Sets a key secret in a keyring
}
