% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{keyring_create}
\alias{keyring_create}
\title{Create a new empty keyring.}
\usage{
keyring_create(keyring, password)
}
\arguments{
\item{keyring}{character(1); Name of keyring}

\item{password}{character(1); Password for keyring}
}
\value{
logical(1); Success or failure of operation
}
\description{
Create a new empty keyring with of a given name with the specified password.
}
\examples{
\dontrun{
keyring_create('mypersonalkeyring', '<PASSWORD>')
}

}
