% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{keyring_exists}
\alias{keyring_exists}
\title{Check if a keyring exists.}
\usage{
keyring_exists(keyring)
}
\arguments{
\item{keyring}{character(1); Name of the keyring.}
}
\value{
logical(1); Keyring file store existence status.
}
\description{
Given a keyring name will check if the keyring file exists.
}
