% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{keyring_lock}
\alias{keyring_lock}
\title{Locks a given keyring}
\usage{
keyring_lock(keyring)
}
\arguments{
\item{character(1);}{Name of keyring}
}
\value{
logical(1); Success or failure of operation
}
\description{
Given the name of a keyring lock it.
}
