% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{keyring_locked}
\alias{keyring_locked}
\title{Is a keyring unlocked for key operations and reading}
\usage{
keyring_locked(keyring)
}
\arguments{
\item{character(1);}{Name of keyring}
}
\value{
logical(1); Success or failure of operation
}
\description{
Query if a keyring is unlocked
}
