% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{keyring_unlock}
\alias{keyring_unlock}
\title{Unlock a keyring.}
\usage{
keyring_unlock(keyring, password)
}
\arguments{
\item{character(1);}{Password for keyring}
}
\value{
logical(1); Success or failure of operation
}
\description{
Unlock a given keyring using the specified password. Secrets exist
in plain text in memory while a keyring is unlocked.
}
