% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsafe_export.R
\name{unsafe_export}
\alias{unsafe_export}
\title{Export keyring to plain text format as a string.}
\usage{
unsafe_export(keyring, format, yaml_tag = "shelter", warn = TRUE)
}
\arguments{
\item{keyring}{character(1); Name of keyring.}

\item{format}{character(1); One of 'yaml' or 'ENV'.}

\item{yaml_tag}{character(1); Tag to use in 'yaml'. Defaults to 'shelter'}

\item{warn}{boolean(1); Should the user be warned of the dangers. Defaults to TRUE.}
}
\value{
A character(1) string of the desired export.
}
\description{
This functions exports a keyring to a file as a convenience function
for production deployments.
}
\details{
WARNING: It is not recommended to use this function
unless you are deploying to a hardened secured production environment. To
restate, if you are developing on a personal laptop a report or code
this function should NOT be used.
}
\examples{
\dontrun{
cat(unsafe_export('mypersonalkeyring', 'yaml'), file="myproject.yml")
}
}
