% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_passwords.R
\docType{data}
\name{common_passwords}
\alias{common_passwords}
\title{A list of common passwords}
\format{
An object of class \code{character} of length 2.
}
\usage{
common_passwords
}
\description{
Dictionary attacks using lists of common passwords are the simplest
most effective attack against any crytpo. Using a password on
this list effectively destroys any possibility of security. In general
do not use any dictionary word without adding some special characters.
}
\keyword{datasets}
