% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{keyring_delete}
\alias{keyring_delete}
\title{Delete a given keyring}
\usage{
keyring_delete(keyring)
}
\arguments{
\item{keyring}{character(1); Name of keyring}
}
\value{
logical(1); Success or failure of operation
}
\description{
Given the name of a keyring, delete it and remove all cached information.
}
\examples{
\dontrun{
keyring_delete('mypersonalkeyring')
}

}
